// src/models/SettingsFormValues.ts
export interface SettingsFormValues {
    boardRows: number;
    boardColumns: number;
    numberOfMinimaxPlayers: number;
    numberOfMinimaxABPlayers: number;
    numberOfExpectimaxPlayers: number;
    numberOfMaxNPlayers: number;
    numberOfUserPlayers: number;
    depth: number;
    timeToThink: number;
}